+++ src/synth-kitchen/modules/gain.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef } from 'react';
 
 import { IAudioContext, IGainNode } from 'standardized-audio-context';
-import { audioContext } from '../audio-context';
+import { audioContext } from '../audio';
 
 import { actions } from '../state/actions';
 import { IModule, IModuleState } from '../state/types/module';
